fix: move associateDtos to correct folder
