+++ backend/src/main/java/com/dbserver/votacao/dto/v1/associate/AssociateDTO.java
@@ -1,4 +1,4 @@
-package com.dbserver.votacao.entity;
+package com.dbserver.votacao.dto.v1.associate;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
